test: UI 테스트와 비즈니스로직 테스트 파일 분리
